requirements.txt
pandas 
gunicorn
numpy
dash==2.11.1
bs4
jupyter_dash
dash_table
dash_bootstrap_components
scikit-learn
plotly.express
base64
joypy
pathlib
plotly.graph_objects